[SPARK-50334][SQL] Extract common logic for reading PB files
